{"schema":"decodex/commit/1","summary":"Split CLI and workflow tests","authority":"manual"}
